+++ src/hybran/run.py
@@ -5,7 +5,7 @@ import subprocess
 from . import extractor, BLAST, CDHIT, MCL, addEggnogAnnotation, parseClustering
 from . import config
 
-def ratt_prokka(ref_dir, fasta, ref_cds, script_dir, cpus, qcov):
+def ratt_prokka(ref_dir, fasta, ref_cds, gcode, script_dir, cpus, qcov):
     """
     Executes RATT and Prokka that resides in ratt_prokka.sh
     File IO is handled by ratt_prokka.sh
@@ -13,6 +13,7 @@ def ratt_prokka(ref_dir, fasta, ref_cds, script_dir, cpus, qcov):
     :param ref_dir: str directory that houses the EMBL reference(s)
     :param fasta: str FASTA file name that needs to be annotated
     :param ref_cds: str FASTA proteome of the reference for Prokka
+    :param gcode: int NCBI genetic code table ID
     :param script_dir: str absolute path to ratt_prokka.sh
     :param cpus: str number of processors/cpus
     :param qcov: int minimum % query coverage (Prokka doesn't have a way of setting ref coverage)
@@ -42,6 +43,7 @@ def ratt_prokka(ref_dir, fasta, ref_cds, script_dir, cpus, qcov):
                ref_cds,
                str(cpus),
                str(qcov),
+               str(gcode),
                ]
         subprocess.call(cmd)
         os.chdir(c)
@@ -97,7 +99,7 @@ def clustering(all_genomes, target_genomes, nproc, seq_ident, seq_covg):
                                             seq_covg=seq_covg)
 
 
-def eggnog_mapper(script_dir, nproc, emapper_loc, temp_dir):
+def eggnog_mapper(script_dir, nproc, emapper_loc, ref_tax_id, ref_gene_dict, temp_dir):
     """
     Runs the run_emapper.sh which executes emapper.py for
     both diamond and hmmer algorithms. The shell script handles
@@ -106,6 +108,8 @@ def eggnog_mapper(script_dir, nproc, emapper_loc, temp_dir):
     :param script_dir: str absolute path to run_emapper.sh
     :param nproc: str number of processors
     :param emapper_loc: str absolute path to eggnog database
+    :param ref_tax_id: str NCBI taxonomy ID
+    :param ref_gene_dict: dict mapping locus tags to gene names
     :param temp_dir: str path to Hybran temporary directory
     :return: None
     """
@@ -120,6 +124,8 @@ def eggnog_mapper(script_dir, nproc, emapper_loc, temp_dir):
     cmd = [os.sep.join([script_dir, 'run_emapper.sh']),
            str(nproc),
            emapper_loc,
-           temp_dir]
+           temp_dir,
+           ref_tax_id,
+           ]
     subprocess.call(cmd)
-    addEggnogAnnotation.update_gbks(script_dir)
+    addEggnogAnnotation.update_gbks(ref_tax_id, ref_gene_dict)